+++ src/test/kotlin/dev/aga/sfm/KReflectionServiceTest.kt
@@ -8,19 +8,17 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
-import org.junit.jupiter.params.provider.Arguments.arguments
 import org.junit.jupiter.params.provider.MethodSource
 import java.lang.reflect.Type
 import kotlin.reflect.KClass
 import kotlin.reflect.javaType
 import kotlin.reflect.typeOf
 
-internal class KInstantiatorFactoryTest {
-
+internal class KReflectionServiceTest {
     @ParameterizedTest
     @MethodSource("testToKClassProvider")
     internal fun testToKClass(target: Type, expected: KClass<*>?) {
-        val actual = KInstantiatorFactory.toKClass(target)
+        val actual = KReflectionService.toKClass(target)
         assertThat(actual).isEqualTo(expected)
     }
 
@@ -29,10 +27,10 @@ internal class KInstantiatorFactoryTest {
         @JvmStatic
         private fun testToKClassProvider(): List<Arguments> {
             return listOf(
-                arguments(typeOf<MyPojo>().javaType, null),
-                arguments(typeOf<KotlinClass>().javaType, null),
-                arguments(typeOf<KotlinObject>().javaType, null),
-                arguments(typeOf<KotlinDataClass>().javaType, KotlinDataClass::class)
+                Arguments.arguments(typeOf<MyPojo>().javaType, null),
+                Arguments.arguments(typeOf<KotlinClass>().javaType, null),
+                Arguments.arguments(typeOf<KotlinObject>().javaType, null),
+                Arguments.arguments(typeOf<KotlinDataClass>().javaType, KotlinDataClass::class)
             )
         }
     }
